refactor: change folder structure for some api methods
