+++ src/globals/api/index.ts
@@ -4,7 +4,7 @@ export * from './methods/catalog-details'
 
 export * from './methods/getAddresses'
 export * from './methods/getLinks'
-export * from './methods/getNewBuildingsOnMain'
+export * from './methods/main-page/getNewBuildingsOnMain'
 export * from './methods/getObjects'
 export * from './methods/getServices'
 export * from './methods/getSocials'
