+++ src/globals/api/methods/catalog-details/getComplexDetailed.ts
