+++ src/globals/api/methods/catalog-details/index.ts
@@ -1,1 +1,2 @@
 export * from './getSecondaryObject'
+export * from './getComplexDetailed'
