+++ src/globals/api/methods/main-page/getNewBuildingsOnMain.ts
